Switched to 'active' flags instead of 'active' TABLES. Untested.
Possibly because of the 'inherited' nature, rows were not being successfully deleted. So, in a slightly unprincipled way, now things simply have an 'active' flag (again) instead of being in an 'active' table. Less efficient, but -- what are the odds we'll ever have more than a few hundred rows to look through anyway, to find the active ones?

Signed-off-by: Alexander Loewi <amloewi@gmail.com>