A new reviewer is searched for from scratch on each rejection (not tested)
Rather than a 'candidates' list there is now a 'rejected' list. In addition, 'assign_reviewer' is its own function, and cuts out most of both 'process_submission' and 'reject_submission' as really it should. Not tested at all though, and there are some big, hastily written changes -- be careful with this one.

Signed-off-by: Alexander Loewi <amloewi@gmail.com>