Sends html emails
The primary purpose was to make 'log on to the PhD Reviewer system' into a link that allowed you to do that ('causa Shelly), and it does does, but now there's some formatting to be done.

I SHOULD really test all of the 'sweep' functionality -- but as soon as I do, the thing's pretty damn complete

Ah fuck, attachments.

Signed-off-by: Alexander Loewi <amloewi@gmail.com>