"disable" disables, miscellaneous improvements, well tested
Previously, it was noticed that a COMPLETED review would actually count as a 'drop'. That was fixed. Also, being disabled actually removes the form for request submission and gives a note why (but not with a countdown till re-enabling on it, which I'd like).
	In addition, a number of small changes were made, such as adding my email at the bottom in case of problems, and the time-aware element of enabling appears to work successfully. You DO get an email when a reviewer chain bottoms out, BUT, two CHAINS can go through the same person. Oh well -- avoiding that would require more coordination that I want to deal with now, and should be unnecessary with more than four reviewers.

Only real pieces remaining: 1) fix the urls, 2) add attachments.

Signed-off-by: Alexander Loewi <amloewi@gmail.com>